Rename module `Typography` to `typography`
