+++ src/imports/components/typography/SectionHeading/index.js
