+++ src/imports/components/typography/SectionHeading/index.styl
