+++ src/imports/components/typography/SectionHeading/mobile/index.styl
