+++ src/imports/components/typography/index.js
@@ -1,0 +1,2 @@
+export {default as Paragraph} from './Paragraph';
+export {default as Heading} from './Heading';
