+++ src/pages/Landing/sections/Services/index.js
@@ -1,5 +1,5 @@
 import React, {PureComponent} from 'react';
-import SectionHeading from 'components/Typography/SectionHeading';
+import SectionHeading from 'components/typography/SectionHeading';
 
 import Service from './components/Service';
 import {Design, Development, Quality} from './services';
