+++ src/pages/Page404/index.js
@@ -2,8 +2,8 @@ import React, {PureComponent} from 'react';
 import {Logo} from 'icons';
 import Loader from 'components/Loader';
 import NavigationBar from 'components/NavigationBar';
-import Heading from 'components/Typography/Heading';
-import Paragraph from 'components/Typography/Paragraph';
+import Heading from 'components/typography/Heading';
+import Paragraph from 'components/typography/Paragraph';
 import Button from 'components/Button';
 import {darkBlue, blue} from 'styles/config';
